pyproject.toml
[tool.poetry]
name = "ideafast-etl"
version = "0.2.0"
description = ""
authors = ["David <davidverweij@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
apache-airflow-providers-mongo = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.15.0"
black = "^21.9b0"
flake8 = "^4.0.1"
flake8-aaa = "^0.12.0"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.3"
flake8-bugbear = "^21.9.2"
isort = "^5.9.3"
mypy = "^0.910"
apache-airflow = "^2.2.0"
coverage = {extras = ["toml"], version = "^6.0.2"}
pytest-cov = "^3.0.0"
nox = "^2021.10.1"
nox-poetry = "^0.8.6"

[tool.poetry.scripts]
build = "cli:build"
version = "cli:version"
bump = "cli:bump"
publish = "cli:publish"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.paths]
source = ["ideafast_etl"]

[tool.coverage.run]
branch = true
source = ["ideafast_etl"]

[tool.coverage.report]
show_missing = true

[tool.isort]
profile = "black"